Rename folder to verification
