+++ packages/core-transactions/__tests__/memoizers/second-signature-verification.test.ts
@@ -1,7 +1,7 @@
 import "jest-extended";
 
 import passphrases from "@packages/core-test-framework/src/internal/passphrases.json";
-import { SecondSignatureVerificationMemoizer } from "@packages/core-transactions/src/memoizers/second-signature-verification";
+import { SecondSignatureVerificationMemoizer } from "@packages/core-transactions/src/verification/second-signature-verification";
 import { Identities, Interfaces, Transactions } from "@packages/crypto";
 import { BuilderFactory } from "@packages/crypto/dist/transactions";
 import { Application, Container, Providers } from "@packages/core-kernel";
